moved deriveSubKeys to helpers
