+++ sources/back-end/libToken/helpers/deriveSubKey.mjs
